refactor: 상품 댓글 목록 CommentList 컴포넌트로 분리
